La tilbake vis vanlig/tom

+++ db/schema.rb
@@ -18,8 +18,8 @@ ActiveRecord::Schema.define(version: 2021_04_12_144538) do
   create_table "active_storage_attachments", force: :cascade do |t|
     t.string "name", null: false
     t.string "record_type", null: false
-    t.integer "record_id", null: false
-    t.integer "blob_id", null: false
+    t.bigint "record_id", null: false
+    t.bigint "blob_id", null: false
     t.datetime "created_at", null: false
     t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
     t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
@@ -37,8 +37,8 @@ ActiveRecord::Schema.define(version: 2021_04_12_144538) do
   end
 
   create_table "allergen_items", force: :cascade do |t|
-    t.integer "allergen_id", null: false
-    t.integer "menu_item_id", null: false
+    t.bigint "allergen_id", null: false
+    t.bigint "menu_item_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["allergen_id"], name: "index_allergen_items_on_allergen_id"
@@ -58,7 +58,7 @@ ActiveRecord::Schema.define(version: 2021_04_12_144538) do
     t.decimal "price"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "menu_id", null: false
+    t.bigint "menu_id", null: false
     t.string "number"
     t.integer "status", default: 1
     t.integer "empty", default: 0
@@ -70,15 +70,15 @@ ActiveRecord::Schema.define(version: 2021_04_12_144538) do
     t.string "description"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "restaurant_id", null: false
+    t.bigint "restaurant_id", null: false
     t.integer "status", default: 1
     t.integer "number"
     t.index ["restaurant_id"], name: "index_menus_on_restaurant_id"
   end
 
   create_table "order_items", force: :cascade do |t|
-    t.integer "menu_item_id", null: false
-    t.integer "order_id", null: false
+    t.bigint "menu_item_id", null: false
+    t.bigint "order_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "quantity", default: 1
@@ -92,7 +92,7 @@ ActiveRecord::Schema.define(version: 2021_04_12_144538) do
     t.text "location"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "restaurant_id", null: false
+    t.bigint "restaurant_id", null: false
     t.index ["restaurant_id"], name: "index_orders_on_restaurant_id"
   end
